Flatten Results hierarchy
