+++ client/src/components/Result/Evaluation/scoreDescriptors.js
