+++ client/src/components/Result/Link.react..test.js
@@ -1,15 +1,15 @@
-import linkProvider from './linkProvider';
+import { getRandomLinks } from './Link';
 
 test('when request three links, return three links.', () => {
   // when
-  const links = linkProvider.getRandomLinks(3);
+  const links = getRandomLinks(3);
 
   // then
   expect(links.length).toBe(3);
 });
 
 test('when requesting more links than present, return all available links', () => {
-  const links = linkProvider.getRandomLinks(1000);
+  const links = getRandomLinks(1000);
 
   const visitedTitles = [];
   links.forEach((link) => {
