+++ client/src/utils/scoreEvaluator.js
